factored RainbowAnimation out of Picker.js
